# Proposed changes
before:
```
mysql [tpch]>select c_comment from customer;
......
15000000 rows in set (9.20 sec)

mysql [tpch]>select trim(c_comment) from customer;
......
15000000 rows in set (10.26 sec)
```

after:
```
mysql [tpch]>select c_comment from customer;
......
15000000 rows in set (9.15 sec)

mysql [tpch]>select trim(c_comment) from customer;
......
15000000 rows in set (9.15 sec)
```

## Problem summary

Introduced a set of low level tools for handwriting simd instructions. Now we can unify our simd usage and easily adapt to greater instructions set.
Considering Doris's main situations to be used are high-performance servers, introduce avx512f(launched in 2016) and avx512bw(launched in 2017)... is meaningful.

## Checklist(Required)

* [ ] Does it affect the original behavior
* [ ] Has unit tests been added
* [ ] Has document been added or modified
* [ ] Does it need to update dependencies
* [x] Is this PR support rollback (If NO, please explain WHY)

## Further comments

Will soonly introduce more tools to completely improve our code's performance. It's expected mainly about string algorithm.